From Bertrand: - Remove "signed/unsigned mismatch" compilation warnings
git-svn-id: http://root.cern.ch/svn/root/trunk@21578 27541ba8-7e3a-0410-8455-c3a389f83636

+++ win32gdk/gdk/src/gdk/gdk.c
@@ -121,7 +121,7 @@ gdk_arg_context_parse(GdkArgContext * context, gint * argc, gchar *** argv)
             break;
          }
 
-         for (j = 0; j < context->tables->len; j++) {
+         for (j = 0; j < (int)context->tables->len; j++) {
             GdkArgDesc *table = context->tables->pdata[j];
             for (k = 0; table[k].name; k++) {
                switch (table[k].type) {
